Removed example of Mozilla's experiments
Both the Grameenphone and Orange examples are not good examples of equal rating but more of a fish and bait tactics. And it really does not bode well with what we are trying to achieve here. 
The idea is good but these two examples do not justify the idea and may cause more harm than good. Have already discussed with Mozilla Privacy telegram group about my concerns.

The additional line I inserted can be neglected if it does not add any value. That was my own opinion;.