Better placeholder text for github issue reporting
Many users just report the email address that has problems and don't
tell us the exact problem. I have changed the placeholder text for the
body of the Github issue, asking people to paste the body of the error
email that they get. It includes the problematic email address, by
default.